Move uses of openToolbox() to showToolbox()
In addition to the conversion to use promises rather than events I
have also removed the host definition unless there was a good reason to
hard code it. This will let us run all the tests with multiple hosts.
Inside gDevTools, openToolbox() as been renamed to _openToolbox() since a
number of functions depend on it still.