Further tweaks to use promises rather than events
Remove setters for Toolbox.target (because it just doesn't work) and
Toolbox.hostType (because it forces you to use events to tell when it
completed) and replace the latter with Toolbox.switchHost.

Make use of host.open returning a promise in Toolbox.open and
Toolbox.switchHost.

Greatly simplify gDevTools.showToolbox using promises from Toolbox as
previously commented.